DJ_A2_01: Renamed routes.ts to routes.constants.ts
