+++ src/routes/index.ts
@@ -1,2 +1,2 @@
-export { ROUTES } from './routes';
+export { ROUTES } from './routes.constants';
 export { router } from './router';
